+++ src/options/options.ts
@@ -1,20 +1,31 @@
 import Alpine from "alpinejs";
+import type { ValuesOf } from "../models/common";
 import type { AlpineOptionsData } from "../models/Alpine";
-import type { Options } from "../models/Options";
+import {
+  COLOR_INDICATOR_OPTIONS,
+  OPTIONS,
+  type Options,
+} from "../models/Options";
 import { type Color, colorsToDots } from "../models/Tabs";
+import { EXTENSION_COMMANDS, type ExtensionCommand } from "../models/Commands";
 import { getOptions } from "../common";
+import {
+  ALPINE_DATA_NAMES,
+  DEFAULT_SHORTCUT,
+  SHORTCUT_KEY_SEPARATOR_REGEXP,
+} from "../constants";
 
 Alpine.data(
-  "options",
+  ALPINE_DATA_NAMES.OPTIONS,
   () =>
     ({
       options: {
-        colorIndicator: "begin",
+        colorIndicator: COLOR_INDICATOR_OPTIONS.BEGIN,
         removeEmptyGroups: true,
         colors: Object.keys(colorsToDots) as Color[],
       },
       colorsToDots,
-      shortcut: ["Ctrl", "+", "Shift", "+", "L"],
+      shortcut: DEFAULT_SHORTCUT,
 
       async init() {
         const options = await getOptions();
@@ -23,15 +34,23 @@ Alpine.data(
         }
 
         const commands = await browser.commands.getAll();
-        this.shortcut = commands
-          .find((command) => command.name === "open-new-tab-in-group")
-          ?.shortcut?.split(/(\+)/g) ?? ["Ctrl", "+", "Shift", "+", "L"];
+        this.shortcut =
+          commands
+            .find(
+              (command) =>
+                command.name === EXTENSION_COMMANDS.OPEN_NEW_TAB_IN_GROUP
+            )
+            ?.shortcut?.split(SHORTCUT_KEY_SEPARATOR_REGEXP) ??
+          DEFAULT_SHORTCUT;
 
         /* eslint-disable-next-line */
         /* @ts-ignore */
         browser.commands.onChanged.addListener((changeInfo) => {
-          if (changeInfo.name === "open-new-tab-in-group") {
-            this.shortcut = changeInfo.newShortcut.split(/(\+)/g);
+          const commandName = changeInfo.name as ExtensionCommand;
+          if (commandName === EXTENSION_COMMANDS.OPEN_NEW_TAB_IN_GROUP) {
+            this.shortcut = changeInfo.newShortcut.split(
+              SHORTCUT_KEY_SEPARATOR_REGEXP
+            );
           }
         });
       },
@@ -42,7 +61,7 @@ Alpine.data(
 
       async setOptions(
         optionName: keyof Options,
-        value: Options[keyof Options]
+        value: ValuesOf<Options>
       ): Promise<void> {
         const newOptions: Options = {
           ...this.options,
@@ -60,7 +79,7 @@ Alpine.data(
           ? this.options.colors.filter((optionColor) => optionColor !== color)
           : [...this.options.colors, color];
 
-        await this.setOptions("colors", newColors);
+        await this.setOptions(OPTIONS.COLORS, newColors);
       },
     }) satisfies AlpineOptionsData
 );